refactor(outline): depluralize outline components representing single instances
