+++ src/app/outline/_components/InteractablesContainer.tsx
@@ -7,15 +7,11 @@ import { MousePointerClick } from 'lucide-react';
 
 import { Card, CardDescription, CardTitle } from '@/components/ui/card';
 import { ScrollArea } from '@/components/ui/scroll-area';
-import {
-  existingOutlineAtom,
-  newOutlineAtom,
-  tutorialOutlineAtom,
-} from '@/lib/atoms';
+import { existingOutlineAtom, newOutlineAtom, tutorialOutlineAtom } from '@/lib/atoms';
 import { cn } from '@/lib/utils';
 
 import TutorialCardComponent from '../tutorial/_components/TutorialCard';
-import Interactables from './Interactables';
+import Interactables from './Interactable';
 
 interface InteractablesContainerProps {
   outlineId: number | null;
